Refactor #1 - sample 데이터 활용할 경우를 위해 타입 분리
